moved all I/Os to lbl_io.py, renamed and cleaned some functions
